Added: GlobalManager.cs class that offloads and refactors the loading and public accessor variable capabilities of the Game1 class. Everything that used to be publicly accessible through the Game1 class is now moved to the GlobalManager class. This allows access to the Main game object itself, the ContentManager, the SpriteBatch, the ShapeDrawer, the GraphicsDeviceMangaer, the ScreenBounds vector, the ScreenCenter vector, the DisplayBounds vector, and a wrapper call to the Random (RNG) property in the InputManager class.
Added: Update, UpdateGUI, Draw, DrawBackground, and DrawGUI methods to the GlobalManager class, that in turn call the appropriate methods in the StateManager and InputManager classes, when GlobalManager is already initialized.

Added: #region and #endregion pre-processor statements and comments that help define areas of the codebase.

Added: A new texture and font loading system based on the usage of common FontIDs and TextureIDs enums, throughout the Asteroids.Tools classes.

Added: Documentation summarizing the functionality of the methods within the GlobalManager class.