Added: Created the PauseState class, that handles functionality during the pause screen.
Refactor: Moved functionality from StateManager into the Pause and State classes. Created a new virtual method, QueueGUIMessages, so that child classes don't have to override the entire UpdateGUI method, in order to draw messages for the user to read.

Change: Made a change to how the debug flag is initialized in classes. Classes on state switch should start out with debug mode set to false.